Fix indentation of generated HTML
Jekyll/Liquid generated HTML has weird indentation that makes it hard
to read.  This patch adds a dependency on a Gem that indent HTML code
everywhere except in `<pre>` tags, where indentation matters, and then
indents the content correctly.

A note on the magic number: the `indent 8` in `default.html` should be
the number of spaces the liquid statement is indented if we want the
indentation to be consistent.