Restructure code for home page
The new Jekyll default is strange: it considers the content of the
home page part of the layout of the site.  This is decidedly not what
we want: the home page is real content, and as such, it should not be
in the layouts.

This patch moves the HTML for the home page into a separate
subdirectory `_pages/`, to keep the root of the repository tidy.  This
means that we need to manually set the permalink of the page, though.
This is in fact a benefit, because we can control our permalinks more
directly, and less incidentally.  Cool URIs never change.