Remove `Gemfile.lock` and fix `.gitignore` (#1)
* Remove `Gemfile.lock`

If you look online, you'll find different opinions about this, mostly
revolving around perceived differences between applications and
gems/libraries.  Regardless, current `Gemfile.lock` will not work on
my system (only half of the gems install, then it gives an error).
This patch removes the broken `Gemfile.lock` and `.gitignore`s it.

* Correct `.gitignore`

The current `.gitignore` file does not ignore several directories that
it looks like it should, and skips the `.bundle/` directory that
stores local Gem information.  The former is fixed by adding trailing
slashes to the names in the `.gitignore`, the latter by adding the
missing ignore.

* Add `_vendor/` to `.gitignore`

This is a good path to install gems locally to this repository,
because it's ignored by Jekyll.