Indentation fixes (#4)
* Fix indentation in `<head>` tag

The initial commit had an indentation of 4 everywhere except the
`<head>` section, where it had an indentation of 2.  @lidavidm says he
wants indentation of 4.

* Fix indentation of generated HTML

Jekyll/Liquid generated HTML has weird indentation that makes it hard
to read.  This patch adds a dependency on a Gem that indent HTML code
everywhere except in `<pre>` tags, where indentation matters, and then
indents the content correctly.

A note on the magic number: the `indent 8` in `default.html` should be
the number of spaces the liquid statement is indented if we want the
indentation to be consistent.