Show future-dated posts by default
This is needed if we want event posts to show up (and if we want
the event post to reflect the event date, not the posting date).
An alternative would be to set `published: true` in the front
matter of every event, but I believe this is cleaner.
Unfortunately, it seems that we can't set this option for only a
certain collection (which would let us limit it to just events).