Generate events iCalendar file
This patch adds a generated iCalendar file at the URL
`/events/calendar.ics` and links to it from the home page and events
page.  The description field isn't always nice looking, but it's the
best I've been able to get it.

The generated iCalendar was validated with [this online
validator][validator].

[validator]: https://icalendar.org/validator.html